auth-fullstack-app-main/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pydantic[email]
python-multipart
fastapi-mail==1.5.0